Initial Submission of Digitized Streaming Example
This is an updated take on the code covered in the DMM6500 Application Note, "Data Logging of Power Profiles from Wireless IoT and Other Low-Power Devices. However, the solution originated on the DMM7510. The Python 3 code extracts the contents of a script file and uploads it to the sockets-connected instrument. The setup, trigger, and data extraction functions are remotely issued from the PC, and we are able to pull data off the DMM at about the same rate that the instrument reading buffer is filling. As the binary data are extracted, they converted to floating point values and then written to a *.csv file.

We noted that the code was written using Python 3, and we did our primary editing and formatting using PyCharm Community Edition to help enforce good programming practices (per helpful suggestions/feedback from the PyCharm IDE).