Control a Temperature Chamber with the DAQ6510
Submission of our work for controlling a temperature chamber enabled with the Watlow F4T controller which offers a LAN connection and SCPI command set. The DAQ6510 script offers up a series of dialogs that allow the operator to define a series of temperature set points and dwell/soak windows. The DAQ then monitors the chamber state to determine when it reaches equilibrium and performs a data logging sequence as defined by the operator. The DAQ repeats the process for each step then ramps the chamber back to a defined safe operating temperature.